requirements.txt
cffi==1.17.1
cryptography==44.0.0
pillow==11.0.0
pycparser==2.22
